ci(versioning): created configuration files to work with the standard-version Node.js tool
Performed research on a suitable automatic versioning and changelog generation tool.
standard-version is currently the best tool for the job given the low number of developers on this
project (currently 1). Regrettably, the tool works best when the project has a package.json file as
that is the only way to pull the repository url for creating links in the changelog.