refactor: created an App struct to encapsulate application lifecycle functions and data
To facilitate testability, readability, and ease of use, the App struct was created to encapsulate
data and behavior related to the application lifecycle. This change enabled the removal of a few Go
files that did not make much sense being on their own. An added benefit is the abstraction of
tview.Application at the main function level. This commit does not build.