test(tools): created a virtual file system to mock directory traversals in tests
The intent of this tool is to limit the static dependency on the built-in filesystem in tests. Basic
commands are simulated and can be injected in directory controller behaviour. An entire filesystem
tree can be easily created with optional seed directory paths that are manually configured. There is
some potential to use the virtual filesystem for manual tests of the program.