refactor: output screen buffer control sequence to stderr instead of stdout
This is a fix for a defect introduced via refactoring. It still classifies as a refactor since the
branch it occurred in has not yet been released. An assumption was made as to the behaviour of the
print() builtin function that was present prior to the refactor. I assumed the function printed to
stdout when it actually printed to stderr. This defect caused the execution scripts to not process
the output of the application correctly and also did not switch the screen buffers.