test: added more list component tests and refactored the virtual filesystem ls command
This change also modifies the way directories display in the list. Directories will no longer have a
trailing slash as they are deemed redundant given we know the list items are all directories. The
trailing slashes contributed to paths being handled inconsistently throughout the code as revealed
via testing. The Ls() function in the virtual filesystem was incomplete and did not support paths
via a parameter, which led to inefficient ways of simulating directory behaviour in tests. The Ls()
function was completed and all tests refactored to pass with the new functionality.