fix: ensured symbolic links are navigable in the directory list
The test for this fix requires the use of the native filesystem in both Windows and Linux. In
Windows, administrator privileges are required to create symbolic links. Tools that would run this
test would need to be opened in administrator mode. The test is skipped if there are insufficient
privileges.

closes #15