feat: introduced different filtering modes for directory list
The filter component has been changed to a form with the original input field, and also a dropdown
with different filtering modes. The default mode is one where the directories are filtered on
beginning text, which in my opinion is the most popular kind of search. There are also ending,
contains, and manual glob pattern filters. The manual glob pattern filter entry is incomplete and
needs refinement, and also needs to ensure consistent operation between Windows and Linux.

closes #12