refactor: ensured DirectoryList creation happens in one function only to reduce parameter repetition
This change was necessary to make now as it would be cumbersome to add parameters otherwise. It did
not make much sense to have two different functions responsible for creating a DirectoryList
instance. The former newDirectoryList() function was a convenience for testing, but not useful for
anything else throughout the application. Any lines of code that were not responsible for creating
data have been moved to an Init() function. After the refactor, the same code runs line by line, but
slightly reordered across CreateDirectoryList() and Init().