fix: ensured consistent directory and file sorting across components and platforms
The weird file ordering encountered in testing was a result of ci simply repeating the order it gets
from the OS. Some directories in Linux have an awkward file ordering that does not display well in
ci and weakens usability. To remedy this, the files and directories are deterministically sorted
before being displayed. The sort algorithm is the same regardless of platform. To accomplish all of
the above, godirwalk has been removed from the project as it no longer serves any purpose in ci,
although it is a good library. Also duplicated the OS path separator constant in test utilities to
avoid circular dependencies.

closes #24