refactor(installation): moved installation scripts to a directory for better project organization
The installation scripts for both platforms were modified so that they can be run from the new
'scripts' directory or the root project directory. This ensures that installation is easy regardless
if it's from a git clone or a release package. During these changes, I realized that it's probably
better if the CodeCov configuration is checked into the repository rather than created dynamically
during the coverage job.