docs(contributing): created the contributing guidelines document
Minor changes have been made as a result of documenting the software development process in the
contributing guidelines. Optimizations have been enabled in production builds of `ci`. The
CODE_OF_CONDUCT.md has been moved to the root repo directory for increased visibility. The LICENSE
file line breaks were adjusted for increased readability. The README.md has been modified with the
correct link to the contributing guidelines document, in addition to adding a badge showcasing this
project's adherence to the Conventional Commits spec.

closes #7